Enable Kotlin 1.3 in the rules_kotlin.
Exposes a toolchain function (define_kt_toolchain) to allow the specification of non-default chains. Sets the default chain to 1.3.21. Removes the now-removed use-coroutines flag, which is permanently on.  This also sets the compiler to permit use of experimental APIs.  This is needed for people migrating from the older experimental coroutines to the new ones, during migration.  This should be configurable, but for now it's just set. Also fixes up some tests and does some minor cleaning.

Fixes bazelbuild/rules_kotlin#159

> Note: This no longer supports using pre-1.3 kotlinc distributions. You can make a toolchain that sets the language and api levels to 1.3, but you must use a 1.3+ kotlinc.