Remove tree artifacts (#287)
remove the generation of unused tree artifacts from the compile step. Tree artifacts, (created by declare_directory) are only useful when an action produces an unknown number of outputs. It this case, this would only result in extra overhead of bundling the produced directories in archives with ever using them.

Co-authored-by: Christian Edward Gruber <christianedwardgruber@gmail.com>