Isolate Kotlin compiler instances for Multiplex workers.
Following the Kotlin Compiler Daemon and conversations with Jetbrains it seems
that we need to set the system property `kotlin.environment.keepalive` so that parallel builds do not clobber each other.

In `exec` the environment is registered as a disposable to be torn down, this system property avoids this.

Switching to a new instance of the compiler for each invocation for safety. There is little state in the compiler and this doesn't affect performance.