Work around stardoc limitations to generate docs
Stardoc doesn't allow a macro to wrap a rule and also have the rule's
documentation be generated.

Ideally, we'd simply exclude the lint wrappers and omit the `input`
attribute from the `stardoc` target. However, thanks to [an issue with
stardoc](https://github.com/bazelbuild/stardoc/issues/99) this is not
an option. As an alternative, we provide a verion of `rules.bzl` that
deliberately excludes the lint wrappers.