Add --exclude_build_data arguemtn to singlejar invocation
Adding --exclude_build_data so that the build-data.properties won't be created.

This file contains contains information about the build host and will therefore break cacheing
for users that are not on the same platform as the cached artifact.

build-data.properties is not created in pure java artifacts so with this change, kotlin artifacts will be inline with java.