Make the underlying _kt library public.
Android kotlin libraries are a macro wrapper, and the underlying kotlin library (for target //foo:bar -> //foo:bar_kt) is marked private. Since the associates= (friendly) internal feature of kotlin requires access to a kotlin library, one is forced to often link test targets to the underlying android _kt backing target, which breaks if you're not in hte same package.

This PR makes it public (which sucks, but kind of is necessary until we can eliminate teh macro-based 'android-sandwich' thing we presently do, so that the associates= feature can work at all in kt_android_* targest.

Also adds an example (which fails without the change) to the android example.

Fixes #540