`ktlint_test` windows support (#785)
* Add vendored windows helpers

This helper is an exact copy of the file
from
https://github.com/aspect-build/bazel-lib/blob/df04df7cd2b9a481787de1cb43a6fe05f67c9511/lib/windows_utils.bzl
Which in turn I believe borrows heavily from
rules_nodejs and bazel internal helpers.

* Add windows support to `ktlint_test`

Unfortunately, the existing shell script
is not launchable by the default windows
CreateProcess command.
See https://bazel.build/rules/windows_tips#actions
for recommendations on how to handle this.
I opted to deal with it in the same way
as rules_nodejs does.

* Optimize depset handling

Taking https://bazel.build/rules/depsets and
https://bazel.build/rules/performance into account,
it is better not to call `to_list()` on a depset.

* Run example - trivial on windows as well

This also ensures both the build as well as
the test are run for windows and linux.

Previously, the tests, and thus the lint
target, was not run. Hence it was not
apparent that the target failed, because
the sample app is formatted incorrectly.

* Apply formatting as per linter

This makes the tests in this
example pass for both windows
and linux

* Bazel commands inherit from one another

So no need to specify options for all
of them.

* Enable runfiles for windows linting

As mentioned in the rule docs.