Bump ktlint to 0.46.1
I have personal motivation to bump this version
due to https://github.com/pinterest/ktlint/issues/1160,
which was release quite a while ago in ktlint already,
but we were still pinned on 0.40.0 here.

Because ktlint can have "breaking" changes as well (new lints, ...)
I think the version _has_ to be configurable. So I did that.