Use Dep.fromLabel in tests to find dependencies path
The Dep.fromLabel works well in combination with kt_rules_test. The same
mangling and demangling is used to find runfiles.

This is a slight simplification, whcih as well prepared the repo for
bzlmod.

In addition workaround for autovalue is replaced with a filegroup trick,
which works with newer rules_jvm_external repository (where visibility
of .jar targets is limited).