Check all dependency jars before marking unused
A dependency might produce more than one jar (e.g. android targets produce a library jar and a resources jar).
If one of those jars is unused, the current logic will mark the dependency as unused. 
This change only marks the dependency unused if none of the jars are used.