[release] Separate bzlmod dev extensions from release extensions
In order to simplify the dependency tree for the rules kotlin bootstrap, bzlmod_setup needed to be separated from the release. The reason for this is twofold: kotlin_repositories in development call kt_configure which configures released_rules_kotlin. Since bzlmod_setup is also used to load released_rules_kotlin this become a dependency cycle. Second, we don't want to load released_rules_koltin in the release module, as it is a development dependency. (whoops.)

This also allows dropping a few extra bits of configuration that are unnecessary for the release.