Fix stardoc breakage in bazel @ HEAD
bazel @ head has an added dependency @rule_java_builtin java_common that is not satisfiable via import the built-in bazel srcs.

Fixing this requires:
* Bump `rules_java` version to one with the toolchain_utils
* Import toolchain_utils from rules_java, not `bazeltools`

Still a potential problem in third_party, as it will pick up toolchain_utils from `bazel_tools` -- unfortunately, the `filegroup` that won't pull in the toolchain_utils is private. :/