Fix android_sdk import path to use rules_android (#1292)
This change updates the `android_sdk` import path in the
`third_party/BUILD` file of `rules_kotlin` to use the `rules_android`
repository instead of the deprecated `bazel_tools` path.
This is necessary for compatibility with Bazel 8.x, as the target
`@bazel_tools//tools/android:android_jar` is no longer usable.