README + Release Note Refresh
I noticed that the `README.md` didn't contain any details about bzlmod support + integration so thought it'd be nice to give it a small refresh.

It looked like the repo integration snippet was also out of date so I decided to point it at the releases page instead like many other rulesets do; however it required adding in some of the extra repo function calls. I updated the release prep script with this and left the existing `WORKSPACE` integration snippet for now until the next release is cut with those added function calls.