Windows portability fixes (#1386)
* buildifier: auto-sorting

* Use native .run action over run_shell

Remove dependency on host shell
since bash is not available on Windows

* Add get_auth to repository_download

In case the content is e.g. behind a mirror with authentication, this is
necessary.

Note: get_auth was introduced in 7.1.0 and would break compatibility
with earlier versions

* Add unsorted-dict-items to the buildifier warnigns

Added fixes with:
```
bazel run //:buildifier.fix
```

* Fix buildifier deprecation

When running `//:buildifier.check` this deprecation warning is issued
```
buildifier: selecting diff program with the BUILDIFIER_DIFF, BUILDIFIER_MULTIDIFF, and DISPLAY environment variables is deprecated, use flags -diff_command and -multi_diff instead
```

* Fix order of sorted flags after _TEST_OPTS got sorted

Since the order of the opts changed, the order in which the flags are
produced changes as well

* Add the unsorted-dict-items checks to ci and pre-commit-config

* Bump buildifier to same version as .bazelci version

bazelci uses 8.2.0